fix: 通过 JEB MCP 并行审查修复代码还原差异
Data/Provider 层修复：
- MotionRecorder: 移除 .clone() 调用和单传感器优化
- Motions: 移除未使用的 MotionComposites/MotionComposite
- Telephony: 移除未使用的 TelephonyComposites/TelephonyComposite
- EmulationMonitorReceiver: WORK_NAME_MONITOR 改为 "emulationMonitor"
- EmulationMonitorWorker: 通知目标改为 EmulateActivity
- Scheduler: 硬编码端口 20230

Map/Extension 层修复：
- Updater: AppUpdater → Updater，产品名改为 "me"
- UpdaterActivity: 更新 Updater 引用
- MainActivity: 更新 Updater 引用
- AppViewModel: 更新 Updater 类型引用
- UpdaterTest: 更新 Updater 引用
- Utility: toFixed 直接格式化 Float

+++ app/src/main/java/com/zhufucdev/motion_emulator/extension/Utility.kt
@@ -226,7 +226,7 @@ fun toFixed(value: Float, n: Int): String {
     }
     val format = DecimalFormat(pattern)
     format.roundingMode = RoundingMode.HALF_UP
-    return format.format(value.toDouble())
+    return format.format(value)
 }
 
 /**
